+++ package.json
@@ -2,7 +2,6 @@
   "name": "howtomakeanoffer",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://romainlp.github.io/howtomakeanoffer/",
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
@@ -15,7 +14,8 @@
     "register-service-worker": "^1.6.2",
     "sass-loader": "^7.1.0",
     "vue": "^2.6.6",
-    "vue-clipboard2": "^0.3.0"
+    "vue-clipboard2": "^0.3.0",
+    "vue-router": "^3.0.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^3.5.0",
@@ -31,6 +31,7 @@
   "gitHooks": {
     "pre-commit": "lint-staged"
   },
+  "homepage": "https://romainlp.github.io/howtomakeanoffer/",
   "lint-staged": {
     "*.{js,vue}": [
       "vue-cli-service lint",
