+++ src/main.js
@@ -2,10 +2,12 @@ import Vue from 'vue'
 import App from './App.vue'
 import './registerServiceWorker'
 import VueClipboard from 'vue-clipboard2'
+import router from './router'
 
 Vue.config.productionTip = false
 Vue.use(VueClipboard);
 
 new Vue({
-  render: h => h(App),
+  router,
+  render: h => h(App)
 }).$mount('#app')
