+++ src/views/Contributors.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="contributors">
+    <h1>Contributors</h1>
+  </div>
+</template>
